transfer one-time transactions

+++ implementation/transaction/src/main/java/com/financetracker/transaction/logic/operations/OneTimeTransactionService.java
@@ -1,9 +1,13 @@
 package com.financetracker.transaction.logic.operations;
 
-import com.financetracker.transaction.api.exceptions.TransactionFailedException;
+import com.financetracker.transaction.api.exceptions.NotParseableException;
+import com.financetracker.transaction.api.exceptions.TransferFailedException;
 import com.financetracker.transaction.infrastructure.db.TransactionRepository;
 import com.financetracker.transaction.infrastructure.client.BankAccountProvider;
 import com.financetracker.transaction.logic.model.OneTimeTransaction;
+import com.financetracker.transaction.logic.model.Transfer;
+import com.financetracker.transaction.logic.model.TransferStatus;
+import jakarta.ws.rs.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +19,7 @@ import java.util.Optional;
 public class OneTimeTransactionService {
 
     private final BankAccountProvider bankAccountProvider;
+    private final TransferService transferService;
     private final TransactionRepository<OneTimeTransaction> oneTimeTransactionRepository;
 
     public List<OneTimeTransaction> getOneTimeTransactions() {
@@ -26,27 +31,58 @@ public class OneTimeTransactionService {
     }
 
     public void createOneTimeTransaction(final OneTimeTransaction oneTimeTransaction) {
-        if (oneTimeTransaction.isInternalTransfer() && !sourceBankAccountBalanceIsSufficientForTransaction(oneTimeTransaction)) {
-            throw new TransactionFailedException();
+        if (sourceAndTargetBankAccountDoNotExist(oneTimeTransaction.getTransfer())) {
+            throw new NotParseableException();
         }
 
-        bankAccountProvider.getBankAccount(oneTimeTransaction.getTransfer().sourceBankAccountId());
         oneTimeTransactionRepository.save(oneTimeTransaction);
-    }
-
-    private boolean sourceBankAccountBalanceIsSufficientForTransaction(final OneTimeTransaction oneTimeTransaction) {
-        final var bankAccount = bankAccountProvider.getBankAccount(oneTimeTransaction.getTransfer().sourceBankAccountId());
-        return bankAccount.filter(bankAccountDto -> bankAccountDto.getBalance() != null && bankAccountDto.getBalance().getAmount() != null && bankAccountDto.getBalance().getAmount() >= oneTimeTransaction.getAmount().amount().doubleValue()).isPresent();
     }
 
     public void updateOneTimeTransaction(final OneTimeTransaction oneTimeTransaction) {
+        if (sourceAndTargetBankAccountDoNotExist(oneTimeTransaction.getTransfer())) {
+            throw new NotParseableException();
+        }
+
+        final Optional<OneTimeTransaction> originalOneTimeTransaction = getOneTimeTransaction(oneTimeTransaction.getId());
+        if (originalOneTimeTransaction.isEmpty()) {
+            throw new NotFoundException();
+        }
+
+        transferService.rollbackTransfer(originalOneTimeTransaction.get().getTransfer(), originalOneTimeTransaction.get());
         oneTimeTransactionRepository.save(oneTimeTransaction);
     }
 
+    private boolean sourceAndTargetBankAccountDoNotExist(final Transfer transfer) {
+        return (transfer.isInternalTransfer() && bankAccountProvider.getBankAccount(transfer.sourceBankAccountId()).isEmpty())
+                || bankAccountProvider.getBankAccount(transfer.targetBankAccountId()).isEmpty();
+    }
 
     public void deleteOneTimeTransaction(final String transactionId) {
+        final Optional<OneTimeTransaction> transaction = getOneTimeTransaction(transactionId);
+
+        if (transaction.isEmpty()) {
+            throw new NotFoundException();
+        }
+
+        transferService.rollbackTransfer(transaction.get().getTransfer(), transaction.get());
         oneTimeTransactionRepository.deleteById(transactionId);
     }
 
+    public void transferOneTimeTransactionAndSetStatus(final String transactionId) {
+        final Optional<OneTimeTransaction> oneTimeTransaction = getOneTimeTransaction(transactionId);
+        if (oneTimeTransaction.isEmpty()) {
+            throw new NotFoundException();
+        }
+
+        try {
+            transferService.transfer(oneTimeTransaction.get().getTransfer(), oneTimeTransaction.get());
+        } catch (TransferFailedException e) {
+            oneTimeTransaction.get().setTransferStatus(TransferStatus.FAILED);
+            oneTimeTransactionRepository.save(oneTimeTransaction.get());
+        }
+
+        oneTimeTransaction.get().setTransferStatus(TransferStatus.SUCCESSFUL);
+        oneTimeTransactionRepository.save(oneTimeTransaction.get());
+    }
 
 }
